[GENERAL] - Start organize directories and code
