+++ app/src/main/AndroidManifest.xml
@@ -5,20 +5,21 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".RickMortyApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.RickMortyApp" >
-        <activity android:name=".ListActivity">
+        android:theme="@style/Theme.RickMortyApp">
+        <activity android:name=".ui.activities.list.ListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".DetailActivity"/>
+        <activity android:name=".ui.activities.detail.DetailActivity" />
 
     </application>
